Refactor bandwidth parsing and limiter into modules
